+++ src/main/java/chess/controller/dto/response/ChessGameScoreResponse.java
@@ -1,4 +1,4 @@
-package chess.controller.dto;
+package chess.controller.dto.response;
 
 import chess.domain.Color;
 import java.util.Map.Entry;
