+++ src/main/java/chess/controller/dto/response/ChessGameStatusResponse.java
@@ -1,4 +1,4 @@
-package chess.controller.dto;
+package chess.controller.dto.response;
 
 public class ChessGameStatusResponse {
